refactor: change app.py to server.py
