+++ packages/send-emails/backend/python/server.py
@@ -4,14 +4,12 @@ import json
 
 from utils.mock_db import db
 
-from io import BytesIO
-from flask import Flask, request, send_file, g
+from flask import Flask, request, g
 from flask_cors import CORS
 
 from nylas import APIClient
 from nylas.client.restful_models import Webhook
 from nylas.services.tunnel import open_webhook_tunnel
-
 
 # Initialize the Nylas API client using the client id and secret specified in the .env file
 nylas = APIClient(
@@ -19,14 +17,11 @@ nylas = APIClient(
     os.environ.get("NYLAS_CLIENT_SECRET"),
 )
 
-# Set the URI for the client
-CLIENT_URI = 'http://localhost:3000'
-
-# Set the default scopes for auth
-DEFAULT_SCOPES = ['email.send', 'email.modify', 'email.read_only']
-
 # Before we start our backend, we should whitelist our frontend
 # as a redirect URI to ensure the auth completes
+#
+# Set the URI for the client
+CLIENT_URI = 'http://localhost:3000'
 updated_application_details = nylas.update_application_details(redirect_uris=[CLIENT_URI])
 print('Application whitelisted. Application Details: ', updated_application_details)
 
@@ -53,8 +48,7 @@ def run_webhook():
 
     open_webhook_tunnel(
         nylas, {'on_message': on_message, 'on_open': on_open, 'on_error': on_error})
-
-
+    
 # Run the webhook
 run_webhook()
 
@@ -71,7 +65,9 @@ def build_auth_url():
     Generates a Nylas Hosted Authentication URL with the given arguments. 
     The endpoint also uses the app level constants CLIENT_URI and DEFAULT_SCOPES to build the URL.
 
-    This endpoint is a POST request and accepts the following parameters in the request body:
+    This endpoint is a POST request and accepts the following parameters:
+    
+    Request Body:
         success_url: The URL to redirect the user to after successful authorization.
         email_address: The email address of the user who is authorizing the app.
 
@@ -84,7 +80,7 @@ def build_auth_url():
     auth_url = nylas.authentication_url(
         (CLIENT_URI or "") + request_body["success_url"],
         login_hint=request_body["email_address"],
-        scopes=DEFAULT_SCOPES,
+        scopes=['email.send', 'email.modify'],
         state=None,
     )
 
@@ -171,81 +167,6 @@ def after_request(response):
     return response
 
 
-@flask_app.route('/nylas/read-emails', methods=['GET'])
-@is_authenticated
-def read_emails():
-    """
-    Retrieve the first 20 threads of the authenticated account from the Nylas API.
-
-    This endpoint is a GET request and accepts no parameters.
-
-    The threads are retrieved using the Nylas API client, with the view set to "expanded".
-
-    The threads are then returned as a JSON object.
-    See our docs for more information about the thread object.
-    https://developer.nylas.com/docs/api/#tag--Threads
-    """
-    res = nylas.threads.where(limit=20, view="expanded").all()
-
-    # enforce_read_only=False is used to return the full thread objects
-    res_json = [item.as_json(enforce_read_only=False) for item in res]
-
-    return res_json
-
-
-@flask_app.route('/nylas/message', methods=['GET'])
-@is_authenticated
-def get_message():
-    """
-    Retrieve a message from the Nylas API.
-
-    This endpoint is a GET request and accepts the following:
-    
-    Query Parameters:
-        'id': The identifier of the message to retrieve.
-
-    The message is retrieved using the Nylas API client by id, with the view set to "expanded".
-
-    The message is then returned as a JSON object.
-    See our docs for more information about the message object.
-    https://developer.nylas.com/docs/api/#tag--Messages
-    """
-
-    message_id = request.args.get('id')
-    message = nylas.messages.where(view="expanded").get(message_id)
-
-    # enforce_read_only=False is required to return the full message object
-    return message.as_json(enforce_read_only=False)
-
-
-@flask_app.route('/nylas/file', methods=['GET'])
-@is_authenticated
-def download_file():
-    """
-    Retrieve and download a file from the Nylas API.
-
-    This endpoint is a GET request and accepts the following:
-
-    Query Parameters:
-        id: The identifier of the file to download.
-
-    The file metadata is retrieved using the Nylas API client.
-    A second request is sent to the API and the file is downloaded.
-
-    Returns the file with metadata to the client for download.
-
-    See our docs for more information about the file object.
-    https://developer.nylas.com/docs/api/#tag--Files
-    """
-
-    file_id = request.args.get('id')
-    file_metadata = nylas.files.get(file_id)
-
-    file = file_metadata.download()
-
-    return send_file(BytesIO(file), download_name=file_metadata.filename, mimetype=file_metadata.content_type, as_attachment=True)
-
-
 @flask_app.route('/nylas/send-email', methods=['POST'])
 @is_authenticated
 def send_email():
@@ -257,9 +178,11 @@ def send_email():
     
     Request Body:
         to: The email address of the recipient.
+        subject: The subject of the email.
         body: The body of the email.
 
     Returns the message object from the Nylas API. 
+    
     See our docs for more information about the message object.
     https://developer.nylas.com/docs/api/#tag--Messages
     """
@@ -281,4 +204,3 @@ def send_email():
 
     # Return the sent message object
     return message
-
